Accept a nil explain on the coordinator; restore reply.c's line endings
SEReply now sends nil for a row without an explanation, so the coordinator's
score reader skips that slot instead of parsing it as an explain node. reply.c
is CRLF on master and an earlier edit had normalised it to LF, inflating the
diff; the endings are restored.

+++ src/aggregate/aggregate_exec.c
@@ -178,7 +178,7 @@ static size_t serializeResult(AREQ *req, RedisModule_Reply *reply, const SearchR
     // One entry per section below, plus the trailing "values" placeholder.
     const size_t entries = !!(options & QEXEC_F_IS_SEARCH) + !!(options & QEXEC_F_SEND_SCORES) +
                            !!(options & QEXEC_F_SENDRAWIDS) + !!(options & QEXEC_F_SEND_PAYLOADS) +
-                           !!(options & QEXEC_F_SEND_SORTKEYS) + need_map + !(options & QEXEC_F_SEND_NOFIELDS) + 1;
+                           !!(options & QEXEC_F_SEND_SORTKEYS) + (size_t)need_map + !(options & QEXEC_F_SEND_NOFIELDS) + 1;
     RedisModule_Reply_MapWithLen(reply, entries);
   }
